requirements.txt
python-telegram-bot[ext]==22.2
motor==3.3.2
APScheduler==3.10.4
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
gigachat>=0.1.0  # For GigaChat integration (optional)
python-dateutil>=2.8.2
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
